Started solution to exercise 3.6.
Also added a shell script to test the solution on every value between 0 and
99 cents, and a Makefile to compile the program and run the test script
(but, students, you *can* do that by typing the commands in the makefile
into a terminal; it is just quicker to type "make it so" than to type
"gcc -Wall blah blah blah").

If you run the test script, you will see that it discovers errors as of
this commit.