Fixed issues with stage 4.
Firstly, Kendall's coefficient requires comparing every pair of rows, not
just consecutive rows. Also, my isconcordant function had an error in it,
as it needed to compare that the values in both columns change in the same
direction between the rows. Previously it was checking the value between
the columns.