Working solution to 9.11
Fixed a bug in the drag function; it would return nan if velocity was zero,
because of a 0/0 error when I tried to get the sign of the velocity. Fixed
by explicitly testing if the velocity is positive or negative.

Also added functions for thrust force, acceleration, and a constant for
gravitational acceleration. Constants also have their units shown next to
them.

Added the rocket simulator to the makefile.